+++ Rdmp.Core/DataExport/DataExtraction/Pipeline/Destinations/ExecuteFullExtractionToDatabaseMSSql.cs
@@ -207,7 +207,6 @@ public class ExecuteFullExtractionToDatabaseMSSql : ExtractionDestination
                         """));
                         return null;
                     }
-
                 }
 
                 if (_request.IsBatchResume)