+++ Rdmp.Core/DataExport/DataExtraction/Pipeline/Destinations/ExecuteFullExtractionToDatabaseMSSql.cs
@@ -267,7 +267,6 @@ public class ExecuteFullExtractionToDatabaseMSSql : ExtractionDestination
                             }
                         }
                     }
-
                 }
 
                 if (_request.IsBatchResume)