+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [9.2.4] - Unreleased
-- Improve SQL Exteraction Merge
+- Improve SQL Extraction Merge
 - Allow Merge code to be viewed in UI
 
 ## [9.2.3] - 2026-05-28