+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [9.2.4] - Unreleased
+- Improve SQL Extraction Merge
+- Allow Merge code to be viewed in UI
 - Fix bug with catalogues not becoming project specific when commit from another project
 - Improve displayed reason for being unable to make Catalogue Project Specific